airflow_examples/teaching_steps/pyproject.toml
[tool.poetry]
name = "teaching-steps"
version = "0.1.0"
description = ""
authors = ["Axel Furlan <axel.furlan@toptal.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "<2.0.0"
requests = "^2.32.3"
numpy = "1.26.4"
redshift-connector = "^2.1.3"
sqlalchemy-redshift = "^0.8.14"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
